New Weapon Spritesheet
Added a spritesheet with 7 new weapons (including the crowbar).

Location: Art > SpaceStation > Interactive > Sprites

Weapons:
-crowbar (new colors)
-pipe wrench
-normal wrench
-broken pipe (2 versions)
-hammer
-mallet (large size)

I created one prefab for the pipe wrench (I just copied the components
from the crowbar prefab so the settings may need to be changed). I
thought the additional weapons could be used to test the loot drop and
inventory systems. (We can add more weapons to the spritesheet in the
future.)