Optimization (minor): spritesheet improvements
I moved the "lightbulb" sprites to the main spritesheet and updated the
lightbulb prefabs. Now all the environment art is in one spritesheet.

Load time and performance have improved a small amount. (I get a minimum
of 200 FPS now.)

NOTE: When making lightbulb prefabs in the future use: lightbulb-01,
lightbulb-02, lightbulb-03 sprites (not the old ones).