Updated SpawnManager and Enemy Behavior
The SpawnManager now has a new variable: “TotalAllowedEnemies” which
sets the maximum number of enemies allowed on screen at one time. This
should also help with performance.

To change the number of enemies go to the gameobject TargetSpawnPoints >
TargetSpawner

I also added some “intelligence” to the enemies. Now they spawn in a
stationary position and only follow the player if the player is within
range (a trigger collider). If the player moves out of range (exits the
trigger collider), the enemy will stop following.