Added (Some) SFX: Player footsteps, Door, LaserWall, Environment
Added audio for the WebGL version (NOT using FMOD)
An animation event calls a function which plays the audio
The SFXs and volume are public variables

-Added open/close SFX to door
-Added player footsteps (walk, run, crouch)
Right now the footstep SFXs for the player's movement are the same, but
the walk, run, and crouch SFX can be separate audio files and have
different volumes. (Crouch is much softer)
-Potential problem: The walk and run SFX are sometimes played very close
together as the animation transitions from walk to run creating a
double-tap sound, so I set the walking volume to 0 (until we come up
with a better fix).
-Added a placeholder for a jetpack SFX (There were no Jetpack audio
files)
-Added "electric" SFX to LaserWall
The LaserWall uses 3D space (so it should be louder the closer you get)
-Added SFX-Environment gameObject (contains WorldSpace sounds: engine
hum, beeps, music, etc)
-The SFX-Environment gameObject should be a good way to test volume
levels
-I couldn't find Hazel's music so I used music from the Unity store as a
placeholder (Assests > Sounds > SFX)

(NOTE: The switches and light on/off code isn't in this build so I
couldn't add the audio for those gameObjects.)